+++ NEWS.md
@@ -14,6 +14,7 @@
     now throw a `TypstCommandError` upon failure
 - `sizehint!(::TypstContext, n)` now returns a `TypstContext` instead of a `Dict{Symbol, Any}`
 - Updated JuliaMono from version 0.55 to 0.61
+- Implement `show_typst` for `Symbol`
 
 ### Bug fixes